[type-stage/viewer] WIP: bring back static document rendering (w/o proseMirror).
