+++ lib/js/components/main-ui.mjs
@@ -61,8 +61,9 @@ export const Layouts = Object.freeze([
   , ['VideoproofArrayDev', 'Videoproof Array (Interim)', VideoproofArrayLayout, LAYOUT_GROUP_DEVELOPMENT]
   , ['Videoproof', 'Video Proof', VideoproofLayout, LAYOUT_GROUP_VIDEO_PROOF]
   , ['TypeToolsGrid', 'Grid', TypeToolsGrid, LAYOUT_GROUP_TYPE_TOOLS]
-  , ['TypeStage', 'Type Stage', TypeStage, LAYOUT_GROUP_DEFAULT]
+
   , ['Ramp', 'Ramp', Ramp, LAYOUT_GROUP_TYPE_TOOLS]
+  , ['TypeStage', 'Type Stage', TypeStage, LAYOUT_GROUP_DEFAULT]
 ]);
 
 /**
